+++ packages/client-common/design/src/constants/index.ts
@@ -1,3 +1,3 @@
 export { colors } from './colors/index.ts';
-export { typography as typo } from './typo/index.ts';
-export type { Typography as Typo } from './typo/index.ts';
+export { typography } from './typography/index.ts';
+export type { Typography } from './typography/index.ts';
