+++ packages/client-common/design/tailwind.config.ts
@@ -1,4 +1,4 @@
-import { colors, typo } from './src/constants/index.ts';
+import { colors, typography } from './src/constants/index.ts';
 import { convertTypoCSS2Tailwind, generatePixelValues } from './src/utils/index.ts';
 
 import type { Config } from 'tailwindcss';
@@ -12,7 +12,7 @@ export default {
 
     colors,
 
-    fontSize: convertTypoCSS2Tailwind(typo),
+    fontSize: convertTypoCSS2Tailwind(typography),
 
     // fontFamily: {
     //   sans: ['Graphik', 'sans-serif'],
